saving current changes to the app, Broken Sections are the following, Timesheet page, and tag page validation in assets section
