+++ shift-scan/src/app/(routes)/admins/assets/@view/jobsite/[id]/page.tsx
@@ -1,22 +1,16 @@
 "use client";
 import { ReusableViewLayout } from "@/app/(routes)/admins/personnel/@view/[employee]/_components/reusableViewLayout";
 import { Holds } from "@/components/(reusable)/holds";
-import { Texts } from "@/components/(reusable)/texts";
-import { Dispatch, SetStateAction, useEffect, useState } from "react";
-import { Inputs } from "@/components/(reusable)/inputs";
-import { Grids } from "@/components/(reusable)/grids";
-import { TextAreas } from "@/components/(reusable)/textareas";
-import { Contents } from "@/components/(reusable)/contents";
-import { Images } from "@/components/(reusable)/images";
-import { Buttons } from "@/components/(reusable)/buttons";
-import { Titles } from "@/components/(reusable)/titles";
-import { EditableFields } from "@/components/(reusable)/EditableField";
+import { useEffect, useState } from "react";
+
 import { deleteAdminJobsite, savejobsiteChanges } from "@/actions/adminActions";
 import { useRouter } from "next/navigation";
-import { Labels } from "@/components/(reusable)/labels";
 import { z } from "zod";
 import { useNotification } from "@/app/context/NotificationContext";
 import { useTranslations } from "next-intl";
+import EditJobsitesHeader from "./_component/jobsiteHeader";
+import { EditJobsitesMain } from "./_component/JobsiteMain";
+import { EditJobsitesFooter } from "./_component/jobsiteFooter";
 
 // Define the Zod schema for Jobsites
 const JobsiteSchema = z.object({
@@ -185,285 +179,12 @@ export default function Jobsites({ params }: { params: { id: string } }) {
   );
 }
 
-//TODO: create edit tags section
-export function EditTags() {
-  const t = useTranslations("Admins");
-  return (
-    <Holds background={"white"} className="w-full h-full col-span-2">
-      <Texts className="text-black font-bold text-2xl">{t("EditTag")}</Texts>
-    </Holds>
-  );
-}
-
-export function EditJobsitesHeader({
-  editedItem,
-  commentText,
-  editFunction,
-  editCommentFunction,
-}: {
-  editedItem: string;
-  commentText: string;
-  editFunction?: Dispatch<SetStateAction<string>>;
-  editCommentFunction?: Dispatch<SetStateAction<string>>;
-}) {
-  const [isCommentSectionOpen, setIsCommentSectionOpen] = useState(false);
-  const t = useTranslations("Admins");
-  const openComment = () => {
-    setIsCommentSectionOpen(!isCommentSectionOpen);
-  };
-  return (
-    <Holds
-      background={"white"}
-      className={`w-full h-full col-span-2 ${
-        isCommentSectionOpen ? "row-span-3" : "row-span-1"
-      }`}
-    >
-      {isCommentSectionOpen ? (
-        <Contents width={"95"} position={"row"}>
-          <Grids
-            rows={"5"}
-            cols={"5"}
-            gap={"2"}
-            className=" h-full w-full py-4"
-          >
-            <Holds className="w-full row-start-1 row-end-2 col-start-1 col-end-4">
-              <Inputs
-                type="text"
-                value={editedItem}
-                onChange={(e) => {
-                  editFunction?.(e.target.value);
-                }}
-                placeholder={t("EditTagName")}
-                variant={"titleFont"}
-                className=" my-auto"
-              />
-            </Holds>
-            <Holds
-              position={"row"}
-              className="h-full w-full my-1 row-start-2 row-end-3 col-start-1 col-end-2 "
-            >
-              <Texts size={"p6"} className="mr-2">
-                {t("Comment")}
-              </Texts>
-              <Images
-                titleImg="/comment.svg"
-                titleImgAlt="comment"
-                size={"30"}
-                onClick={openComment}
-                className="cursor-pointer hover:shadow-black hover:shadow-md"
-              />
-            </Holds>
-
-            <Holds className="w-full h-full row-start-3 row-end-6 col-start-1 col-end-6">
-              <TextAreas
-                placeholder={t("EnterYourComment")}
-                value={commentText ? commentText : ""}
-                onChange={(e) => {
-                  editCommentFunction?.(e.target.value);
-                }}
-                maxLength={40}
-                rows={4}
-                style={{ resize: "none" }}
-              />
-            </Holds>
-          </Grids>
-        </Contents>
-      ) : (
-        <Grids rows={"2"} cols={"5"} className=" h-full w-full p-4">
-          <Holds className="w-full row-start-1 row-end-3 col-start-1 col-end-4 ">
-            <Inputs
-              type="text"
-              value={editedItem}
-              placeholder={t("EditTagName")}
-              onChange={(e) => {
-                editFunction?.(e.target.value);
-              }}
-              variant={"titleFont"}
-              className=" my-auto"
-            />
-          </Holds>
-          <Holds
-            position={"row"}
-            className="h-full w-full row-start-2 row-end-3 col-start-6 col-end-7 "
-          >
-            <Texts size={"p6"} className="mr-2">
-              {t("Comment")}
-            </Texts>
-            <Images
-              titleImg="/comment.svg"
-              titleImgAlt="comment"
-              size={"40"}
-              onClick={openComment}
-              className="cursor-pointer hover:shadow-black hover:shadow-md"
-            />
-          </Holds>
-        </Grids>
-      )}
-    </Holds>
-  );
-}
-export function EditJobsitesMain({
-  formState,
-  handleFieldChange,
-  hasChanged,
-  originalState,
-}: {
-  formState: {
-    name: string;
-    streetName: string;
-    streetNumber: string;
-    city: string;
-    state: string;
-    country: string;
-    description: string;
-    comment: string;
-  };
-  handleFieldChange: (field: string, value: string) => void;
-  hasChanged: (
-    field:
-      | "name"
-      | "streetName"
-      | "streetNumber"
-      | "city"
-      | "state"
-      | "country"
-      | "description"
-      | "comment"
-  ) => boolean;
-  originalState: {
-    name: string;
-    streetName: string;
-    streetNumber: string;
-    city: string;
-    state: string;
-    country: string;
-    description: string;
-    comment: string;
-  };
-}) {
-  const t = useTranslations("Admins");
-  return (
-    <Holds background={"white"} className="w-full h-full ">
-      <Grids cols={"3"} rows={"3"} gap={"5"} className="w-full h-full p-4  ">
-        {/* Input */}
-        <Holds className="h-full w-full ">
-          <Labels size={"p6"}>
-            {t("StreetName")} <span className="text-red-500">*</span>
-          </Labels>
-          <EditableFields
-            value={formState.streetName}
-            isChanged={hasChanged("streetName")}
-            onChange={(e) => handleFieldChange("streetName", e.target.value)}
-            onRevert={() =>
-              handleFieldChange("streetName", originalState.streetName)
-            }
-            variant="default"
-            size="default"
-          />
-        </Holds>
-        <Holds className="h-full w-full">
-          <Labels size={"p6"}>
-            {t("StreetNumber")} <span className="text-red-500">*</span>
-          </Labels>
-          <EditableFields
-            value={formState.streetNumber}
-            isChanged={hasChanged("streetNumber")}
-            onChange={(e) => handleFieldChange("streetNumber", e.target.value)}
-            onRevert={() =>
-              handleFieldChange("streetNumber", originalState.streetNumber)
-            }
-            variant="default"
-            size="default"
-          />
-        </Holds>
-        <Holds className="h-full w-full">
-          <Labels size={"p6"}>
-            {t("City")} <span className="text-red-500">*</span>
-          </Labels>
-          <EditableFields
-            value={formState.city}
-            isChanged={hasChanged("city")}
-            onChange={(e) => handleFieldChange("city", e.target.value)}
-            onRevert={() => handleFieldChange("city", originalState.city)}
-            variant="default"
-            size="default"
-          />
-        </Holds>
-        <Holds className="h-full w-full">
-          <Labels size={"p6"}>
-            {t("State")} <span className="text-red-500">*</span>
-          </Labels>
-          <EditableFields
-            value={formState.state}
-            isChanged={hasChanged("state")}
-            onChange={(e) => handleFieldChange("state", e.target.value)}
-            onRevert={() => handleFieldChange("state", originalState.state)}
-            variant="default"
-            size="default"
-          />
-        </Holds>
-        <Holds className="h-full w-full">
-          <Labels size={"p6"}>
-            {t("Country")} <span className="text-red-500">*</span>
-          </Labels>
-          <EditableFields
-            value={formState.country}
-            isChanged={hasChanged("country")}
-            onChange={(e) => handleFieldChange("country", e.target.value)}
-            onRevert={() => handleFieldChange("country", originalState.country)}
-            variant="default"
-            size="default"
-          />
-        </Holds>
-        <Holds className="h-full w-full col-start-1 col-end-4 row-start-3 row-end-4 ">
-          <Labels size={"p6"}>
-            {t("Description")} <span className="text-red-500">*</span>
-          </Labels>
-          <TextAreas
-            value={formState.description}
-            onChange={(e) => handleFieldChange("description", e.target.value)}
-          />
-        </Holds>
-      </Grids>
-    </Holds>
-  );
-}
-
-export function EditJobsitesFooter({
-  handleEditForm,
-  deleteTag,
-}: {
-  handleEditForm: () => void;
-  deleteTag: () => void;
-}) {
-  const t = useTranslations("Admins");
-  return (
-    <Holds
-      background={"white"}
-      className="w-full h-full row-span-1 col-span-2 "
-    >
-      <Grids cols={"4"} gap={"4"} className="w-full h-full p-4">
-        <Holds className=" col-start-1 col-end-2  ">
-          <Buttons
-            background={"red"}
-            className="py-2"
-            onClick={() => {
-              deleteTag();
-            }}
-          >
-            <Titles size={"h4"}>{t("DeleteJobsite")}</Titles>
-          </Buttons>
-        </Holds>
-
-        <Holds className="col-start-4 col-end-5 ">
-          <Buttons
-            className={"py-2 bg-app-green"}
-            onClick={() => handleEditForm()}
-          >
-            <Titles size={"h4"}>{t("SubmitEdit")}</Titles>
-          </Buttons>
-        </Holds>
-      </Grids>
-    </Holds>
-  );
-}
+// //TODO: create edit tags section
+// export function EditTags() {
+//   const t = useTranslations("Admins");
+//   return (
+//     <Holds background={"white"} className="w-full h-full col-span-2">
+//       <Texts className="text-black font-bold text-2xl">{t("EditTag")}</Texts>
+//     </Holds>
+//   );
+// }
