+++ shift-scan/src/app/(routes)/admins/assets/@view/jobsite/page.tsx
@@ -2,7 +2,7 @@ import { Holds } from "@/components/(reusable)/holds";
 import EmptyView from "../../../_pages/EmptyView";
 import { useTranslations } from "next-intl";
 
-export default function jobsitemain() {
+export default function Jobsitemain() {
   const t = useTranslations("Admins");
   return (
     <Holds className="w-full h-full ">
