+++ shift-scan/src/app/(routes)/admins/assets/@view/new-tag/page.tsx
@@ -1,28 +1,24 @@
 "use client";
 import { createTag } from "@/actions/adminActions";
 import { ReusableViewLayout } from "@/app/(routes)/admins/personnel/@view/[employee]/_components/reusableViewLayout";
-import { CheckBox } from "@/components/(inputs)/checkBox";
-import { Buttons } from "@/components/(reusable)/buttons";
-import { Contents } from "@/components/(reusable)/contents";
-import { Grids } from "@/components/(reusable)/grids";
 import { Holds } from "@/components/(reusable)/holds";
-import { Images } from "@/components/(reusable)/images";
-import { Inputs } from "@/components/(reusable)/inputs";
-import { TextAreas } from "@/components/(reusable)/textareas";
-import { Texts } from "@/components/(reusable)/texts";
-import { Titles } from "@/components/(reusable)/titles";
 import { costCodesTag, JobTags } from "@/lib/types";
-import { useTranslations } from "next-intl";
-import { Dispatch, SetStateAction, useEffect, useMemo, useState } from "react";
+import { useEffect, useState } from "react";
 import { z } from "zod";
 import { useNotification } from "@/app/context/NotificationContext";
+import NewTagHeader from "./_Components/NewTagHeader";
+import { NewTagMainRight } from "./_Components/NewTagRightMain";
+import NewTagFooter from "./_Components/NewTagFooter";
+import NewTagMainLeft from "./_Components/NewTagLeftMain";
 
 // Zod schema for validation
 const tagPayloadSchema = z.object({
   name: z.string().min(1, "Name is required"),
   description: z.string().optional(),
   jobs: z.array(z.string()).nonempty("At least one job must be selected"),
-  costCodes: z.array(z.string()).nonempty("At least one cost code must be selected"),
+  costCodes: z
+    .array(z.string())
+    .nonempty("At least one cost code must be selected"),
 });
 
 export default function NewTagView() {
@@ -126,7 +122,7 @@ export default function NewTagView() {
       <ReusableViewLayout
         custom={true}
         header={
-          <EditTagHeader
+          <NewTagHeader
             editedItem={editedItem}
             commentText={commentText}
             editFunction={setEditedItem}
@@ -135,7 +131,7 @@ export default function NewTagView() {
         }
         mainHolds="h-full w-full flex flex-row row-span-6 col-span-2 bg-app-dark-blue px-4 py-2 rounded-[10px] gap-4"
         mainLeft={
-          <EditTagMainLeft
+          <NewTagMainLeft
             toggleJobSelection={toggleJobSelection}
             toggleCostCodeSelection={toggleCostCodeSelection}
             jobs={jobs ?? []}
@@ -145,317 +141,13 @@ export default function NewTagView() {
           />
         }
         mainRight={
-          <EditTagMainRight
+          <NewTagMainRight
             selectedJobs={selectedJobs}
             selectedCostCodes={selectedCostCodes}
           />
         }
-        footer={<EditTagFooter handleEditForm={handleCreateTag} />}
+        footer={<NewTagFooter handleEditForm={handleCreateTag} />}
       />
-    </Holds>
-  );
-}
-
-export function EditTagHeader({
-  editedItem,
-  commentText,
-  editFunction,
-  editCommentFunction,
-}: {
-  editedItem: string;
-  commentText: string;
-  editFunction?: Dispatch<SetStateAction<string>>;
-  editCommentFunction?: Dispatch<SetStateAction<string>>;
-}) {
-  const [isCommentSectionOpen, setIsCommentSectionOpen] = useState(false);
-  const t = useTranslations("Admins");
-  const openComment = () => {
-    setIsCommentSectionOpen(!isCommentSectionOpen);
-  };
-  return (
-    <Holds
-      background={"white"}
-      className={`w-full h-full col-span-2 ${
-        isCommentSectionOpen ? "row-span-3" : "row-span-1"
-      }`}
-    >
-      {isCommentSectionOpen ? (
-        <Contents width={"95"} position={"row"}>
-          <Grids
-            rows={"5"}
-            cols={"5"}
-            gap={"2"}
-            className=" h-full w-full py-4"
-          >
-            <Holds className="w-full row-start-1 row-end-2 col-start-1 col-end-4">
-              <Inputs
-                type="text"
-                value={editedItem}
-                onChange={(e) => {
-                  editFunction?.(e.target.value);
-                }}
-                placeholder={t("EditTagName")}
-                variant={"titleFont"}
-                className=" my-auto"
-              />
-            </Holds>
-            <Holds
-              position={"row"}
-              className="h-full w-full my-1 row-start-2 row-end-3 col-start-1 col-end-2 "
-            >
-              <Texts size={"p6"} className="mr-2">
-                {t("Comment")}
-              </Texts>
-              <Images
-                titleImg="/comment.svg"
-                titleImgAlt="comment"
-                size={"30"}
-                onClick={openComment}
-                className="cursor-pointer hover:shadow-black hover:shadow-md"
-              />
-            </Holds>
-
-            <Holds className="w-full h-full row-start-3 row-end-6 col-start-1 col-end-6">
-              <TextAreas
-                placeholder={t("EnterYourComment")}
-                value={commentText ? commentText : ""}
-                onChange={(e) => {
-                  editCommentFunction?.(e.target.value);
-                }}
-                maxLength={40}
-                rows={4}
-                style={{ resize: "none" }}
-              />
-            </Holds>
-          </Grids>
-        </Contents>
-      ) : (
-        <Grids rows={"2"} cols={"5"} className=" h-full w-full p-4">
-          <Holds className="w-full row-start-1 row-end-3 col-start-1 col-end-4 ">
-            <Inputs
-              type="text"
-              value={editedItem}
-              placeholder={t("EditTagName")}
-              onChange={(e) => {
-                editFunction?.(e.target.value);
-              }}
-              variant={"titleFont"}
-              className=" my-auto"
-            />
-          </Holds>
-          <Holds
-            position={"row"}
-            className="h-full w-full row-start-2 row-end-3 col-start-6 col-end-7 "
-          >
-            <Texts size={"p6"} className="mr-2">
-              {t("Comment")}
-            </Texts>
-            <Images
-              titleImg="/comment.svg"
-              titleImgAlt="comment"
-              size={"40"}
-              onClick={openComment}
-              className="cursor-pointer hover:shadow-black hover:shadow-md"
-            />
-          </Holds>
-        </Grids>
-      )}
-    </Holds>
-  );
-}
-
-export function EditTagMainLeft({
-  toggleJobSelection,
-  toggleCostCodeSelection,
-  jobs,
-  costCodes,
-  selectedJobs,
-  selectedCostCodes,
-}: {
-  toggleJobSelection: (job: JobTags) => void;
-  toggleCostCodeSelection: (costCode: costCodesTag) => void;
-  jobs: JobTags[];
-  costCodes: costCodesTag[];
-  selectedJobs: JobTags[];
-  selectedCostCodes: costCodesTag[];
-}) {
-  const t = useTranslations("Admins");
-  const [term, setTerm] = useState<string>("");
-  const [activeList, setActiveList] = useState<"jobs" | "costCodes">("jobs");
-
-  const filteredCostCodes = useMemo(() => {
-    return costCodes.filter(
-      (cc) =>
-        cc.description.toLowerCase().includes(term.toLowerCase()) ||
-        cc.name.toLowerCase().includes(term.toLowerCase())
-    );
-  }, [term, costCodes]);
-
-  const filteredJobs = useMemo(() => {
-    return jobs.filter((tag) =>
-      tag.name.toLowerCase().includes(term.toLowerCase())
-    );
-  }, [term, jobs]);
-  return (
-    <Holds background="white" className="w-full h-full p-4">
-      <Grids rows="10" gap="5" className="h-full">
-        <Holds
-          position={"row"}
-          background={"white"}
-          className="row-span-2 h-full gap-4 border-[3px] rounded-[15px] border-black"
-        >
-          <Holds
-            background={activeList === "jobs" ? "lightBlue" : "white"}
-            className="w-[50%] h-full justify-center "
-            onClick={() => setActiveList("jobs")}
-          >
-            <Texts size={"p6"}>{t("Jobsite")}</Texts>
-          </Holds>
-          <Holds
-            background={activeList === "costCodes" ? "lightBlue" : "white"}
-            className="w-[50%] h-full justify-center "
-            onClick={() => setActiveList("costCodes")}
-          >
-            <Texts size={"p6"}>{t("CostCode")}</Texts>
-          </Holds>
-        </Holds>
-
-        <Holds className="row-span-8 h-full border-[3px] border-black rounded-t-[10px]">
-          <Holds position="row" className="py-2 border-b-[3px] border-black">
-            <Holds className="h-full w-[20%]">
-              <Images titleImg="/magnifyingGlass.svg" titleImgAlt="search" />
-            </Holds>
-            <Holds className="w-[80%]">
-              <Inputs
-                type="search"
-                placeholder={t("Search")}
-                value={term}
-                onChange={(e) => setTerm(e.target.value)}
-                className="border-none outline-none"
-              />
-            </Holds>
-          </Holds>
-
-          {activeList === "costCodes" ? (
-            <Holds className="h-full mb-4 overflow-y-auto no-scrollbar">
-              {filteredCostCodes.map((costCode) => (
-                <Holds
-                  key={costCode.id}
-                  className="py-2 border-b cursor-pointer flex items-center"
-                >
-                  <Holds position={"row"} className="justify-between">
-                    <Holds className="flex w-2/3">
-                      <Texts size="p6">{`${costCode.name} ${costCode.description}`}</Texts>
-                    </Holds>
-                    <Holds position="row" className="relative flex w-1/3">
-                      <CheckBox
-                        id={costCode.id.toString()}
-                        checked={selectedCostCodes.some(
-                          (c) => c.id === costCode.id
-                        )}
-                        onChange={() => toggleCostCodeSelection(costCode)}
-                        size={2}
-                        name={costCode.name}
-                        aria-label={`Toggle tag ${costCode.name}`}
-                      />
-                    </Holds>
-                  </Holds>
-                </Holds>
-              ))}
-            </Holds>
-          ) : (
-            <Holds className="h-full mb-4 overflow-y-auto no-scrollbar">
-              {filteredJobs.map((job) => (
-                <Holds
-                  key={job.id}
-                  className="py-2 border-b cursor-pointer flex items-center"
-                >
-                  <Holds position={"row"} className="justify-between">
-                    <Holds className="flex w-2/3">
-                      <Texts size="p6">{job.name}</Texts>
-                    </Holds>
-                    <Holds position="row" className="relative flex w-1/3">
-                      <CheckBox
-                        id={job.id.toString()}
-                        checked={selectedJobs.some((j) => j.id === job.id)}
-                        onChange={() => toggleJobSelection(job)}
-                        size={2}
-                        name={job.name}
-                        aria-label={`Toggle tag ${job.name}`}
-                      />
-                    </Holds>
-                  </Holds>
-                </Holds>
-              ))}
-            </Holds>
-          )}
-        </Holds>
-      </Grids>
-    </Holds>
-  );
-}
-
-export function EditTagMainRight({
-  selectedJobs,
-  selectedCostCodes,
-}: {
-  selectedJobs: JobTags[];
-  selectedCostCodes: costCodesTag[];
-}) {
-  return (
-    <Holds background={"white"} className="w-full h-full">
-      <Grids rows={"2"} gap={"2"} className="w-full h-full p-2">
-        <Holds className="row-start-1 row-end-2 h-full bg-slate-200 rounded-[10px] overflow-y-auto no-scrollbar">
-          {/* Flex container with flex-wrap */}
-          <Holds className="flex flex-row flex-wrap gap-2 w-full p-2 ">
-            {selectedJobs.map((item, index) => (
-              <Holds
-                key={index}
-                className="w-fit h-fit p-1 bg-white border-[3px] border-black rounded-[10px] "
-              >
-                <Texts size={"p6"}>{item.name}</Texts>
-              </Holds>
-            ))}
-          </Holds>
-        </Holds>
-        <Holds className="row-start-2 row-end-3 h-full bg-slate-200 rounded-[10px] overflow-y-auto no-scrollbar">
-          {/* Flex container with flex-wrap */}
-          <Holds className="flex flex-row flex-wrap gap-2 w-full p-2  ">
-            {selectedCostCodes.map((item, index) => (
-              <Holds
-                key={index}
-                className="w-fit h-fit p-1 bg-white border-[3px] border-black rounded-[10px] "
-              >
-                <Texts size={"p6"}>{item.description}</Texts>
-              </Holds>
-            ))}
-          </Holds>
-        </Holds>
-      </Grids>
-    </Holds>
-  );
-}
-export function EditTagFooter({
-  handleEditForm,
-}: {
-  handleEditForm: () => void;
-}) {
-  const t = useTranslations("Admins");
-  return (
-    <Holds
-      background={"white"}
-      className="w-full h-full row-span-1 col-span-2 "
-    >
-      <Grids cols={"4"} gap={"4"} className="w-full h-full p-4">
-        <Holds className="col-start-4 col-end-5 ">
-          <Buttons
-            className={"py-2 bg-app-green"}
-            onClick={() => handleEditForm()}
-          >
-            <Titles size={"h4"}>{t("CreateTag")}</Titles>
-          </Buttons>
-        </Holds>
-      </Grids>
     </Holds>
   );
 }
