+++ shift-scan/src/app/(routes)/admins/personnel/@view/timesheets/[employee]/page.tsx
@@ -38,7 +38,7 @@ export default function Timesheets({
       setLoading(false);
     };
     fetchUserName();
-  }, [params.employee]);
+  }, [params.employee, t]);
   return (
     <Holds className="h-full w-full">
       {loading ? (
