+++ shift-scan/src/data/dataValues.ts
@@ -1,5 +1,3 @@
-import CostCodes from "@/app/(routes)/admin/assets/(components)/costcodes";
-import Jobsite from "@/app/(routes)/admin/assets/(components)/jobsite";
 import { Prisma } from "@prisma/client";
 
 export const now = new Date();
